+++ .packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2021-01-12 21:35:37.672904.
+# Generated by pub on 2021-01-12 22:35:10.355291.
 _fe_analyzer_shared:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/_fe_analyzer_shared-12.0.0/lib/
 analyzer:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/analyzer-0.40.6/lib/
 animated_splash_screen:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/animated_splash_screen-1.0.1+2/lib/
@@ -125,6 +125,7 @@ string_scanner:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pu
 synchronized:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/synchronized-2.2.0+2/lib/
 term_glyph:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0-nullsafety.3/lib/
 test_api:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.2.19-nullsafety.6/lib/
+timeline_tile:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/timeline_tile-1.0.0/lib/
 transparent_image:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/transparent_image-1.0.0/lib/
 tuple:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/tuple-1.0.3/lib/
 typed_data:file:///Users/kazuosaito/development/flutter/.pub-cache/hosted/pub.dartlang.org/typed_data-1.3.0-nullsafety.5/lib/
