+++ pubspec.yaml
@@ -47,7 +47,7 @@ dependencies:
   location: ^3.1.0
   firebase_storage: ^5.2.0
   animated_splash_screen: ^1.0.1+2
-
+  timeline_tile: ^1.0.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
